tests/openresult.test.ts
import { handleKeyboardEvent } from '../src/openresult';

describe('handleKeyboardEvent', () => {
  let originalLocationSearch, originalWindowOpen;

  beforeEach(() => {
    // Store the original values to restore them later
    originalLocationSearch = window.location.search;
    originalWindowOpen = window.open;

    // Mock the relevant properties and methods
    Object.defineProperty(window.location, 'search', {
      writable: true,
      value: '',
    });
    window.open = jest.fn();
  });

  afterEach(() => {
    // Restore the original values
    window.location.search = originalLocationSearch;
    window.open = originalWindowOpen;
  });
  
  it('should click on the first product card', () => {
    // Create a mock KeyboardEvent object
    const event = new KeyboardEvent('keydown', {
      key: '1',
      ctrlKey: false,
      altKey: false, // Add more properties as needed
    });

    // Simulate the keypress event with the mock KeyboardEvent
    handleKeyboardEvent(event);

    // Perform assertions to check the expected behavior
    expect(window.location.search).toBe(''); // Check if the location.search was not modified
    expect(window.open).not.toHaveBeenCalled(); // Ensure that window.open was not called
    // Add more assertions based on your specific HTML structure and expectations
  });

  it('should open the first product card in a new tab when Ctrl is held', () => {
    // Create a mock KeyboardEvent object
    const event = new KeyboardEvent('keydown', {
      key: '1',
      ctrlKey: true,
      altKey: false, // Add more properties as needed
    });

    // Simulate the keypress event with the mock KeyboardEvent
    handleKeyboardEvent(event);

    // Perform assertions to check the expected behavior
    expect(window.location.search).toBe(''); // Check if the location.search was not modified
    expect(window.open).toHaveBeenCalled(); // Ensure that window.open was called
    // Add more assertions based on your specific expectations
  });
});